fix: split lineage file
